Introduced tree module.
